search well-known locations for metadata xml files
Depending on the init script, jira may start with a working
directory in $jira.home/bin.  So, let's check the parent if
we couldn't find the xml file in the current directory.

Also, if using catalina (the default), check first in
catalina.base, since this should point to the JIRA install
location regardless of the init script.

Finally, log the locations we checked in the system log so
the administrator stands a chance of finding where to put
the files.